backend/requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.36
alembic==1.12.1
psycopg2-binary==2.9.9

# Email
aiosmtplib==3.0.1

# OpenAI
openai==1.3.7

# Twilio
twilio==8.10.0
resend==0.8.0

# HTTP Requests
httpx==0.25.1
requests==2.31.0

# Data Processing
pandas==2.3.3
python-dateutil==2.9.0

# Environment
python-dotenv==1.0.0

# CORS
python-jose[cryptography]==3.3.0

# Validation
pydantic==2.9.2
pydantic-settings==2.6.1

# Web Scraping (for research)
beautifulsoup4==4.12.2
selenium==4.15.2

# Playwright (for dynamic content scraping)
playwright==1.40.0

# Firecrawl (for API-based scraping)
firecrawl-py==0.0.16

# Sentiment Analysis
textblob==0.17.1
transformers==4.46.2

# LinkedIn (unofficial)
#linkedin-api==2.2.0

# Calendar
#icalendar==5.0.11

# Excel Export
openpyxl==3.1.2

# Logging
loguru==0.7.2

# Async File I/O (latency optimization)
aiofiles==23.2.1

# Background Tasks
#celery==5.3.4
#redis==5.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1